Move finite-state-machine to core/data-structures
